pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "hexmap"
version = "0.0.0"
description = "Terminal-first binary explorer with schema overlay (Textual)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Contributors" }]
dependencies = [
  "textual>=0.45",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.4.5",
]

[project.urls]
Homepage = "https://example.com/hexmap"
Repository = "https://example.com/hexmap/repo"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = []

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

